refactored common modal logic out of signup modal
