+++ client/src/components/organisms/CommonModal.tsx
@@ -1,0 +1,34 @@
+import React, { useState } from 'react';
+import { useApolloClient } from "@apollo/client";
+import { useDispatch } from "react-redux";
+
+import { Button, Modal } from 'react-bootstrap';
+
+function CommonModal(props: any) {
+    const [show, setShow] = useState(true);
+
+    const handleClose = () => setShow(false);
+    const handleShow = () => setShow(true);
+  
+    return ( 
+            <Modal show={show} onHide={handleClose}>
+                <Modal.Header closeButton>
+                <Modal.Title>{props.modalTitle}</Modal.Title>
+                </Modal.Header>
+                <Modal.Body>
+                    {props.body}
+                </Modal.Body>
+                <Modal.Footer>
+                <Button variant="secondary" onClick={handleClose}>
+                    Close
+                </Button>
+                <Button variant="primary" onClick={handleClose}>
+                    Save Changes
+                </Button>
+                </Modal.Footer>
+            </Modal>
+    );
+  }
+  
+export default CommonModal;
+
