--- client/src/pages/SignUp.tsx
@@ -1,52 +1,0 @@
-import React, { useState } from 'react';
-import { useApolloClient } from "@apollo/client";
-import { useDispatch } from "react-redux";
-
-import { Button, Modal } from 'react-bootstrap';
-
-function SignUp() {
-    
-    const client = useApolloClient();
-    const [email, setEmail] = useState("");
-    const [password, setPassword] = useState("");
-    const [firstName, setFirstName] = useState("");
-    const [lastName, setLastName] = useState("");
-    const dispatch = useDispatch();
-  
-    const handleClick = async (e: React.ChangeEvent<HTMLInputElement>) => { 
-        e.preventDefault();
-        //dispatch(register(firstName, lastName, email, password, client ));
-    };
-
-    const [show, setShow] = useState(false);
-
-    const handleClose = () => setShow(false);
-    const handleShow = () => setShow(true);
-
-    const onChangeFirstName = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setFirstName(e.target.value);
-    };
-
-    const onChangeLastName = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setLastName(e.target.value);
-    }
-
-    const onChangeEmail = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setEmail(e.target.value);
-    };
-
-    const onChangePass = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setPassword(e.target.value);
-    }
-  
-    return (
-        <div>
-
-            
-        </div>
-            
-    );
-  }
-  
-export default SignUp;
-
